feat: delete plants via DEL
