+++ frontend/src/features/map_planning/layers/plant/components/PlantLayerLeftToolbar.tsx
@@ -1,9 +1,6 @@
 import { useIsReadOnlyMode } from '../../../utils/ReadOnlyModeContext';
-import {
-  DeletePlantAction,
-  UpdateAddDatePlantAction,
-  UpdateRemoveDatePlantAction,
-} from '../actions';
+import { UpdateAddDatePlantAction, UpdateRemoveDatePlantAction } from '../actions';
+import { useDeleteSelectedPlantings } from '../hooks/useDeleteSelectedPlantings';
 import {
   PlantingDateAttribute,
   SinglePlantingAttributeForm,
@@ -16,9 +13,8 @@ export function PlantLayerLeftToolbar() {
     (state) => state.untrackedState.layers.plants.selectedPlantings,
   );
   const executeAction = useMapStore((state) => state.executeAction);
-  const selectPlantings = useMapStore((state) => state.selectPlantings);
-  const transformerRef = useMapStore((state) => state.transformer);
   const step = useMapStore((state) => state.step);
+  const { deleteSelectedPlantings } = useDeleteSelectedPlantings();
 
   const isReadOnlyMode = useIsReadOnlyMode();
 
@@ -42,14 +38,7 @@ export function PlantLayerLeftToolbar() {
   };
 
   const onDeleteClick = () => {
-    if (!selectedPlantings?.length) return;
-
-    selectedPlantings.forEach((selectedPlanting) =>
-      executeAction(new DeletePlantAction({ id: selectedPlanting.id })),
-    );
-
-    selectPlantings(null);
-    transformerRef.current?.nodes([]);
+    deleteSelectedPlantings();
   };
 
   if (nothingSelected) {
