+++ frontend/src/features/map_planning/utils/layer-utils.ts
@@ -1,0 +1,11 @@
+import useMapStore from '../store/MapStore';
+
+export function isPlantLayerActive(): boolean {
+  const selectedLayerType = useMapStore.getState().getSelectedLayerType();
+  return selectedLayerType === 'plants';
+}
+
+export function isBaseLayerActive(): boolean {
+  const selectedLayerType = useMapStore.getState().getSelectedLayerType();
+  return selectedLayerType === 'base';
+}
